AI-Chat-with-Camera-Integration/requirements.txt
streamlit>=1.31.0
opencv-python>=4.9.0
Pillow>=10.0.0
anthropic
python-dotenv>=1.0.0
websockets>=11.0
elevenlabs
pydub
groq